third_party/blink/web_tests/fast/dom/HTMLLinkElement/link-preload-unsupported-destination-expected.txt

CONSOLE WARNING: <link rel=preload> must have a valid `as` value
CONSOLE WARNING: <link rel=preload> uses an unsupported `as` value
CONSOLE WARNING: <link rel=preload> uses an unsupported `as` value
CONSOLE WARNING: <link rel=preload> uses an unsupported `as` value
